\name{coef.bivar}
\alias{coef.bivar}
\alias{coefficients.bivar}
\title{Extract the Coefficients for the Bivariate Symbolic Regression Method}
\description{Returns the coefficients from an object class \code{bivar}.}
\usage{
\method{coef}{bivar}(object, ...)
}
\arguments{
  \item{object}{ an object class \code{bivar}. }
  \item{\dots}{ other arguments. }
}
\value{ Coefficients extracted from an object class \code{bivar}. }

\seealso{   \code{\link{bivar}}}

\keyword{regression}