\name{coef.ccrm}
\alias{coef.ccrm}
\alias{coefficients.ccrm}
\title{Extract the Coefficients for the Constrained Centre and Range Method}
\description{Returns the coefficients from an object class \code{ccrm}. }
\usage{
\method{coef}{ccrm}(object, ...)
}
\arguments{
  \item{object}{ an object class \code{ccrm}. }
  \item{\dots}{ other arguments. }
}
\value{ Coefficients extracted from an object class \code{object}. }

\seealso{   \code{\link{ccrm}}}

\keyword{regression}