\name{coef.crm}
\alias{coef.crm}
\alias{coefficients.crm}
\title{Extract the Coefficients for the Centre and Range Method}
\description{Returns the coefficients from an object class \code{crm}. }
\usage{
\method{coef}{crm}(object, ...)
}
\arguments{
  \item{object}{ an object class \code{crm}. }
  \item{\dots}{ other arguments. }
}
\value{ Coefficients extracted from an object class \code{object}. }

\seealso{   \code{\link{crm}}}

\keyword{regression}