\name{coef.MinMax}
\alias{coef.MinMax}
\alias{coefficients.MinMax}
\title{Extract MinMax Method Coefficients}
\description{Returns the coefficients from an object class \code{MinMax}.}
\usage{
\method{coef}{MinMax}(object, ...)
}
\arguments{
  \item{object}{ an object class \code{MinMax}. }
  \item{\dots}{ other arguments. }
}
\value{ Coefficients extracted from an object class \code{MinMax}. }

\seealso{   \code{\link{MinMax}}}

\keyword{regression}