\name{fitted.bivar}
\alias{fitted.bivar}
\alias{fitted.values.bivar}
\title{Extract Bivariate Symbolic Regression Method Fitted Values}
\description{Returns the fitted values from an object class \code{bivar}.}
\usage{
\method{fitted}{bivar}(object, ...)
}
\arguments{
  \item{object}{ an object class \code{bivar}. }
  \item{\dots}{ other arguments. }
}
\value{ Fitted values extracted from the object class \code{bivar}. }

\seealso{   \code{\link{bivar}}}

\keyword{regression}