\name{fitted.ccrm}
\alias{fitted.ccrm}
\alias{fitted.values.ccrm}
\title{Extract Constrained Centre and Range Method Fitted Values}
\description{Returns the fitted values from an object class \code{ccrm}. }
\usage{
\method{fitted}{ccrm}(object, ...)
}
\arguments{
  \item{object}{ an object class \code{ccrm}. }
  \item{\dots}{ other arguments. }
}
\value{ Fitted values extracted from the object class \code{object}. }

\seealso{   \code{\link{ccrm}}}

\keyword{regression}