\name{fitted.icm}
\alias{fitted.icm}
\alias{fitted.values.icm}
\title{Extract Centre Method Fitted Values}
\description{Returns the fitted values from an object class \code{icm}. }
\usage{
\method{fitted}{icm}(object, ...)
}
\arguments{
  \item{object}{ an object class \code{icm}. }
  \item{\dots}{ other arguments. }
}
\value{ Fitted values extracted from an object class \code{icm}. }

\seealso{   \code{\link{icm}}}

\keyword{regression}