\name{fitted.crm}
\alias{fitted.crm}
\alias{fitted.values.crm}
\title{Extract Centre and Range Method Fitted Values}
\description{Returns the fitted values from an object class \code{crm}. }
\usage{
\method{fitted}{crm}(object, ...)
}
\arguments{
  \item{object}{ an object class \code{crm}. }
  \item{\dots}{ other arguments. }
}
\value{ Fitted values extracted from the object class \code{object}. }

\seealso{   \code{\link{crm}}}

\keyword{regression}