\name{fitted.MinMax}
\alias{fitted.MinMax}
\alias{fitted.values.MinMax}
\title{Extract MinMax Method Fitted Values}
\description{Returns the fitted values from an object class \code{MinMax}.}
\usage{
\method{fitted}{MinMax}(object, ...)
}
\arguments{
  \item{object}{ an object class \code{MinMax}. }
  \item{\dots}{ other arguments. }
}
\value{ Fitted values extracted from the object class \code{MinMax}. }

\seealso{   \code{\link{MinMax}}}

\keyword{regression}