\name{modm}
\alias{modm}

\title{}
\description{ \code{modm} is ...}
\usage{
modm(xmin, xup, ymin, yup)
}
\arguments{
  \item{xmin}{...}
  \item{xup}{...}
  \item{ymin}{...}
  \item{yup}{...}  
}
\details{
...
}
\value{
	\code{Ymin_est}{...}
	\item{Ymax_est}{...}
	\item{alpha}{...}
	\item{beta}{beta}
	\item{Bmin}{...}
	\item{Bmax}{...}
	\item{gamma}{...}
	\item{delta}{...}

}
\references{

Blanco Fernandez et. al 2011

}
\author{Eufrasio de A. Lima Neto \email{eufrasio@de.ufpb.br}, Pedro Rafael D. Marinho \email{pedro.rafael.marinho@gmail.com}}

\note{...}

\seealso{
\code{\link{summary.ccrm}}, \code{\link{coef.ccrm}}, \code{\link{fitted.ccrm}}, \code{\link{residuals.ccrm}}, \code{\link{formula}}
}

\examples{...}
\keyword{modm}

