\name{print-iRegression}
\alias{print.icm}
\alias{print.crm}
\alias{print.ccrm}
\alias{print.MinMax}
\alias{print.bivar}
\alias{print.summary.icm}
\alias{print.summary.crm}
\alias{print.summary.ccrm}
\alias{print.summary.MinMax}
\alias{print.summary.bivar}
\alias{print.coef.crm}
\alias{print.coef.ccrm}
\alias{print.coef.MinMax}
\alias{print.coef.bivar}
\title{Print Values for various iRegression methods}
\description{\code{print} prints its argument. }
\usage{
\method{print}{icm}(x, ...)
\method{print}{crm}(x, ...)
\method{print}{ccrm}(x, ...)
\method{print}{MinMax}(x, ...)
\method{print}{bivar}(x, ...)
\method{print}{summary.icm}(x, ...)
\method{print}{summary.crm}(x, ...)
\method{print}{summary.ccrm}(x, ...)
\method{print}{summary.MinMax}(x, ...)
\method{print}{summary.bivar}(x, ...)
\method{print}{coef.crm}(x, ...)
\method{print}{coef.ccrm}(x, ...)
\method{print}{coef.MinMax}(x, ...)
\method{print}{coef.bivar}(x, ...)
}
\arguments{
  \item{x}{ an object used to select a method.. }
  \item{\dots}{ further arguments passed to or from other methods. }
}

\seealso{   \code{\link{print}}}

\keyword{print}