\name{residuals.bivar}
\alias{residuals.bivar}
\title{Extract Bivariate Symbolic Regression Method Residuals}
\description{Returns the residuals from an object class \code{bivar}.}
\usage{
\method{residuals}{bivar}(object, ...)
}
\arguments{
  \item{object}{ an object class \code{bivar}. }
  \item{\dots}{ other arguments. }
}

\value{ Residuals extracted from the object class \code{bivar}. }

\seealso{   \code{\link{bivar}}}

\keyword{regression}