\name{residuals.ccrm}
\alias{residuals.ccrm}
\title{Extract Constrained Centre and Range Method Residuals}
\description{Returns the residuals from an object class \code{ccrm}.}
\usage{
\method{residuals}{ccrm}(object, ...)
}
\arguments{
  \item{object}{ an object class \code{ccrm}. }
  \item{\dots}{ other arguments. }
}
\value{ Residuals extracted from the object class \code{ccrm}. }

\seealso{   
\code{\link{ccrm}}
}

\keyword{regression}