\name{residuals.cm}
\alias{residuals.cm}
\title{Extract Centre Method Residuals}
\description{Returns the residuals from an object class \code{cm}. }
\usage{
\method{residuals}{cm}(object, ...)
}
\arguments{
  \item{object}{ an object class \code{cm}. }
  \item{\dots}{ other arguments. }
}
\value{ Residuals extracted from the object class \code{cm}. }

\seealso{   \code{\link{cm}}}

\keyword{regression}