\name{residuals.crm}
\alias{residuals.crm}
\title{Extract Centre and Range Method Residuals}
\description{Returns the residuals from an object class \code{crm}.}
\usage{
\method{residuals}{crm}(object, ...)
}
\arguments{
  \item{object}{ an object class \code{crm}. }
  \item{\dots}{ other arguments. }
}
\value{ Residuals extracted from the object class \code{crm}. }

\seealso{   
\code{\link{crm}}
}

\keyword{regression}