\name{residuals.MinMax}
\alias{residuals.MinMax}
\title{Extract MinMax Method Residuals}
\description{Returns the residuals from an object class \code{MinMax}.}
\usage{
\method{residuals}{MinMax}(object, ...)
}
\arguments{
  \item{object}{ an object class \code{MinMax}. }
  \item{\dots}{ other arguments. }
}

\value{ Residuals extracted from the object class \code{MinMax}. }

\seealso{   \code{\link{MinMax}}}

\keyword{regression}