\name{fitted.cm}
\alias{fitted.cm}
\alias{fitted.values.cm}
\title{Extract Centre Method Fitted Values}
\description{Returns the fitted values from an object class \code{cm}. }
\usage{
\method{fitted}{cm}(object, ...)
}
\arguments{
  \item{object}{ an object class \code{cm}. }
  \item{\dots}{ other arguments. }
}
\value{ Fitted values extracted from an object class \code{cm}. }

\seealso{   \code{\link{cm}}}

\keyword{regression}