\name{iRegression-package}
\alias{iRegression-package}
\alias{iRegression}

\docType{package}
\title{ iRegression Package }
\description{

The \code{iRegression} package is used to fit regression methods for symbolic interval-valued variables. At this time, the package provide the
following linear methods: Centre Method (Billard and Diday, 2000), MinMax Method and Centre and Range Method (Lima Neto and De Carvalho, 2008).

}
\details{
\tabular{ll}{
Package: \tab iRegression\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2011-06-01\cr
License: \tab GPL (>= 2) \cr
LazyLoad: \tab yes\cr
}

Some available functions: 
\code{\link{cm}}, \code{\link{MinMax}}, \code{\link{crm}}}

\author{
Eufrasio de A. Lima Neto \email{eufrasio@de.ufpb.br} and Claudio A. V. de Souza Filho \email{cldsza@gmail.com}

Maintainer: Eufrasio de A. Lima Neto \email{eufrasio@de.ufpb.br}

Acknowledgments: The authors would like to thank CNPq (Brazilian Agency) for their financial support.
}

\references{

Billard, L. and Diday, E. (2000) Regression analysis for interval-valued data. \emph{Data Analysis, Classification
and Related Methods: Proceedings of the Seventh Conference of the International
Federation of Classification Societies}, Springer-Verlag, pp. 369-374.

Lima Neto, E.A. and De Carvalho, F.A.T. (2008). Centre and range method to fitting a linear regression model on
symbolic interval data. \emph{Computational Statistics and Data Analysis}, \bold{52}, 1500--1515.

}
