% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modm.R
\name{modm}
\alias{modm}
\title{Model M}
\usage{
modm(formula1, formula2, data, ...)
}
\arguments{
\item{formula_min}{formula com o minimo.}

\item{formula_max}{formula com o maximo.}
}
\value{
\itemize{
\item Ymin_est
\item Ymax_est
\item alpha
\item beta
\item Bmin
\item Bmax
\item gamma
\item delta
}
}
\description{
Alguma descricao a respeito do modelo M.

Descrevendo o pacote.
}
\details{
Algum detalhe do modelo e/ou uso da funcao.
}
\references{
Alguma referencia do Modelo M.
}
\seealso{
\code{\link{cm}}, \code{\link{ccrm}} and \code{\link{MinMax}}.
}
\author{
Pedro Rafael D. Marinho and Eufrasio de Andrade Lima Neto.
}
