\name{print-iRegression}
\alias{print.cm}
\alias{print.crm}
\alias{print.MinMax}
\alias{print.summary.cm}
\alias{print.summary.crm}
\alias{print.summary.MinMax}
\alias{print.coef.crm}
\alias{print.coef.MinMax}
\title{Print Values for various iRegression methods}
\description{\code{print} prints its argument. }
\usage{
\method{print}{cm}(x, ...)
\method{print}{crm}(x, ...)
\method{print}{MinMax}(x, ...)
\method{print}{summary.cm}(x, ...)
\method{print}{summary.crm}(x, ...)
\method{print}{summary.MinMax}(x, ...)
\method{print}{coef.crm}(x, ...)
\method{print}{coef.MinMax}(x, ...)
}
\arguments{
  \item{x}{ an object used to select a method.. }
  \item{\dots}{ further arguments passed to or from other methods. }
}

\seealso{   \code{\link{print}}}

\keyword{print}