\name{residuals.icm}
\alias{residuals.icm}
\title{Extract Centre Method Residuals}
\description{Returns the residuals from an object class \code{icm}. }
\usage{
\method{residuals}{icm}(object, ...)
}
\arguments{
  \item{object}{ an object class \code{icm}. }
  \item{\dots}{ other arguments. }
}
\value{ Residuals extracted from the object class \code{icm}. }

\seealso{   \code{\link{icm}}}

\keyword{regression}